Crate triemap

Structs

DrainIter

A draining iterator over the key-value pairs of a TrieMap.

IntoIter

An owning iterator over the key-value pairs of a TrieMap.

IntoKeys

An owning iterator over the keys of a TrieMap.

IntoValues

An owning iterator over the values of a TrieMap.

Iter

An iterator over the key-value pairs of a TrieMap.

IterMut

A mutable iterator over the key-value pairs of a TrieMap.

Keys

An iterator over the keys of a TrieMap.

OccupiedEntry

A view into an occupied entry in a TrieMap.

PrefixIter

An iterator over entries with keys that start with a specific prefix.

PrefixKeys

Iterator for keys that start with a specific prefix.

PrefixValues

Iterator for values that have keys starting with a specific prefix.

TrieMap

A TrieMap is a key-value data structure that uses a trie (prefix tree) for storage and retrieval of data.

VacantEntry

A view into a vacant entry in a TrieMap.

Values

An iterator over the values of a TrieMap.

ValuesMut

A mutable iterator over the values of a TrieMap.

Enums

Entry

Represents an entry in a TrieMap which may either be vacant or occupied.

Traits

AsBytes

The AsBytes trait allows a type to be used as a key in a TrieMap.